+++ src/main/java/com/shop/JewleryMS/controller/Test.java
@@ -1,8 +1,7 @@
-package com.shop.JewleryMS.Controller;
+package com.shop.JewleryMS.controller;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
