+++ src/main/java/com/shop/JewleryMS/model/CreateCategoryRequest.java
@@ -1,17 +1,9 @@
-package com.shop.JewleryMS.Entity;
+package com.shop.JewleryMS.model;
 
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.Setter;
+import lombok.Data;
 
-@Entity
-@Getter
-@Setter
-@Table(name = "Category")
-public class Category {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    long id;
+@Data
+public class CreateCategoryRequest {
     String name;
     String description;
