+++ src/main/java/com/shop/JewleryMS/repository/AuthenticationRepository.java
@@ -1,10 +1,8 @@
 package com.shop.JewleryMS.repository;
 
-import com.shop.JewleryMS.Entity.Account;
+import com.shop.JewleryMS.entity.Account;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
-import java.util.List;
 
 @Repository
 public interface AuthenticationRepository extends JpaRepository<Account, Long> {
