+++ src/main/java/com/shop/JewleryMS/repository/CategoryRepository.java
@@ -1,6 +1,6 @@
 package com.shop.JewleryMS.repository;
 
-import com.shop.JewleryMS.Entity.Category;
+import com.shop.JewleryMS.entity.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
