+++ src/main/java/com/shop/JewleryMS/repository/CustomerRepository.java
@@ -1,6 +1,6 @@
 package com.shop.JewleryMS.repository;
 
-import com.shop.JewleryMS.Entity.Customer;
+import com.shop.JewleryMS.entity.Customer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CustomerRepository extends JpaRepository<Customer, Long> {
