+++ src/main/java/com/shop/JewleryMS/repository/ProductSellRepository.java
@@ -1,6 +1,6 @@
 package com.shop.JewleryMS.repository;
 
-import com.shop.JewleryMS.Entity.ProductSell;
+import com.shop.JewleryMS.entity.ProductSell;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
